requirements.txt
# Last update: Today
# Project: Rag systems evaluation pipeline using Ragas
# Authored by TII/AICCU/Edge-Team
# Final version: Implemented multi-system support (RagFlow, Dify) and improved API handling.


aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.0.0
deepdiff==8.0.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
effdet==0.4.1
emoji==2.12.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
filelock==3.16.0
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
greenlet==3.1.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.7
humanfriendly==10.0
idna==3.10
iopath==0.1.10
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.2.13
langchain-community==0.2.12
langchain-core==0.2.40
langchain-openai==0.1.21
langchain-text-splitters==0.2.4
langdetect==1.0.9
langsmith==0.1.121
layoutparser==0.3.4
lxml==5.3.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
newsapi-python==0.2.7
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
olefile==0.47
omegaconf==2.3.0
onnx==1.16.2
onnxruntime==1.19.2
openai==1.40.6
opencv-python==4.10.0.84
openpyxl==3.1.5
orderly-set==5.2.2
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pi_heif==0.18.0
pikepdf==9.2.1
pillow==10.4.0
portalocker==2.10.1
proto-plus==1.24.0
protobuf==5.28.1
psutil==6.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pyparsing==3.1.4
pypdf==5.0.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-oxmsg==0.0.1
pytz==2024.2
PyYAML==6.0.2
ragas==0.1.13
rapidfuzz==3.9.7
regex==2024.9.11
reportlab==4.2.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==3.0.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
sympy==1.13.2
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.9
tokenizers==0.19.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.15.12
unstructured-client==0.25.8
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.13
urllib3==2.2.3
wrapt==1.16.0
xmltodict==0.13.0
xxhash==3.5.0
yarl==1.11.1